rename internals
