+++ bfabric/src/bfabric/experimental/entity_lookup_cache.py
@@ -6,7 +6,7 @@ from typing import TypeVar, TYPE_CHECKING
 
 from loguru import logger
 
-from bfabric.experimental.cache.fifo_cache import FifoCache
+from bfabric.experimental.cache._fifo_cache import FifoCache
 
 if TYPE_CHECKING:
     from collections.abc import Generator
